Move components in common folder
