+++ frontend/app/common/components/SigninLogo.tsx
